Move content checker factory to where the checkers live
